refactor(test): reorganize test helpers and utils
