+++ test/utils/crypto.ts
@@ -1,13 +1,11 @@
 import { ethers } from 'hardhat';
-import type { CoreRequestTypes, CoreResultTypes, ProverDataTypes } from '../ts-types';
-
-export const SEDA_DATA_TYPES_VERSION = '0.0.1';
-export const ONE_DAY_IN_SECONDS = 24 * 60 * 60;
-
-const RESULT_DOMAIN_SEPARATOR = '0x00';
-const SECP256K1_DOMAIN_SEPARATOR = '0x01';
-
-const NON_ZERO_HASH = ethers.keccak256(ethers.toUtf8Bytes('0x'));
+import type { CoreRequestTypes, CoreResultTypes, ProverDataTypes } from '../../ts-types';
+import {
+  NON_ZERO_HASH,
+  RESULT_DOMAIN_SEPARATOR,
+  SECP256K1_DOMAIN_SEPARATOR,
+  SEDA_DATA_TYPES_VERSION,
+} from './constants';
 
 function padBigIntToBytes(value: bigint, byteLength: number): string {
   return ethers.zeroPadValue(ethers.toBeArray(value), byteLength);
